RF: Added messages provider
