+++ src/utils/scrolling.ts
@@ -7,6 +7,9 @@
  * @exports scrollToBottom - Function to scroll to the bottom of an element
  */
 
+// Internal Modules
+import { DivRefObject } from "../types/aliases";
+
 
 
 /**
@@ -14,7 +17,7 @@
  * @param elementRef - The React element to scroll to the bottom of
  * @param smooth - Whether the scroll is gradual or instant
  */
-export function scrollToBottom(elementRef: React.RefObject<HTMLDivElement>, smooth: boolean = false) {
+export function scrollToBottom(elementRef: DivRefObject, smooth: boolean = false) {
   // Access the current value of the Ref
   const element = elementRef.current;
   if (!element) { return; }
